Permite agregar o eliminar objetivos
Objective reporter ahora permite agregar y eliminar objetivos

+++ Assets/Scripts/Systems/ObjectiveReporter.cs
@@ -2,25 +2,24 @@ using UnityEngine;
 
 public class ObjectiveReporter : MonoBehaviour
 {
+    [Header("Objetivo a agregar")]
+    [SerializeField] private bool _addObjective;
     [SerializeField] private string _objectiveMessage;
     [SerializeField] private ObjectiveManager.ObjectiveId _objectiveId;
 
     [Header("Objetivo a eliminar")]
-    [SerializeField] private bool deleteObjective;
+    [SerializeField] private bool _deleteObjective;
     [SerializeField] private ObjectiveManager.ObjectiveId _objectiveIdRemove;
 
-    private bool objetiveChanged;
+    private bool objectiveChanged;
 
     public void ReportObjective()
     {
-        if (objetiveChanged) return;
+        if (objectiveChanged) return;
 
-        ObjectiveManager.Instance.UpdateObjective(_objectiveId, _objectiveMessage);
-        if (deleteObjective)
-        {
-            ObjectiveManager.Instance.DeleteObjective(_objectiveIdRemove);
-        }
+        if (_addObjective) ObjectiveManager.Instance.UpdateObjective(_objectiveId, _objectiveMessage);
+        if (_deleteObjective) ObjectiveManager.Instance.DeleteObjective(_objectiveIdRemove);
 
-        objetiveChanged = true;
+        objectiveChanged = true;
     }
 }
